Allow replacement of Jenkins config with configMap. (#7450)
Currently, the configMap is only used to write the initial
config.xml, etc to /var/jenkins_home.  Any changes to the
config are not used by Jenkins, because the `cp` operation
uses `--no-clobber`.

This adds a new `.Values.Master.OverwriteConfig` option to
allow a user to specify that the configMap should always
overwrite the config.xml, jenkins.CLI.xml, and
jenkins.model.JenkinsLocationConfiguration.xml
files.

By default, the `.Values.Master.OverwriteConfig` option is `false`
to preserve the current behavior.

Fixes #3133.

Signed-off-by: Logan Owen <logan@s1network.com>